frame and test for REUC exstension (#293)
