+++ git-index/src/extension/decode.rs
@@ -37,6 +37,9 @@ pub fn all(maybe_beginning_of_extensions: &[u8], object_hash: git_hash::Kind) ->
             extension::tree::SIGNATURE => {
                 ext.tree = extension::tree::decode(ext_data, object_hash);
             }
+            extension::resolve_undo::SIGNATURE => {
+                ext.resolve_undo = extension::resolve_undo::decode(ext_data, object_hash);
+            }
             extension::end_of_index_entry::SIGNATURE => {} // skip already done
             extension::index_entry_offset_table::SIGNATURE => {} // not relevant/obtained already
             mandatory if mandatory[0].is_ascii_lowercase() => match mandatory {
@@ -60,5 +63,6 @@ pub fn all(maybe_beginning_of_extensions: &[u8], object_hash: git_hash::Kind) ->
 pub struct Outcome {
     pub tree: Option<extension::Tree>,
     pub link: Option<extension::Link>,
+    pub resolve_undo: Option<extension::resolve_undo::Paths>,
     pub is_sparse: bool,
 }
