+++ git-index/src/extension/mod.rs
@@ -35,67 +35,42 @@ pub(crate) mod end_of_index_entry;
 
 pub(crate) mod index_entry_offset_table;
 
+pub mod link;
+
+pub(crate) mod resolve_undo {
+    use crate::extension::Signature;
+    use bstr::BString;
+    use git_hash::ObjectId;
+
+    pub type Paths = Vec<ResolvePath>;
+
+    pub struct ResolvePath {
+        /// relative to the root of the repository, or what would be stored in the index
+        name: BString,
+
+        /// common ancestor, stage 1
+        common: Option<Stage>,
+        /// stage 2
+        ours: Option<Stage>,
+        /// stage 3
+        theirs: Option<Stage>,
+    }
+
+    pub struct Stage {
+        mode: u32,
+        id: ObjectId,
+    }
+
+    pub const SIGNATURE: Signature = *b"REUC";
+
+    pub fn decode(data: &[u8], object_hash: git_hash::Kind) -> Option<Paths> {
+        todo!("decode REUC")
+    }
+}
+
 pub mod sparse {
     use crate::extension::Signature;
 
     /// Only used as an indicator
     pub const SIGNATURE: Signature = *b"sdir";
-}
-
-pub mod link {
-    use crate::extension::{Link, Signature};
-    use crate::util::split_at_pos;
-
-    pub const SIGNATURE: Signature = *b"link";
-
-    pub struct Bitmaps {
-        pub delete: git_bitmap::ewah::Vec,
-        pub replace: git_bitmap::ewah::Vec,
-    }
-
-    pub mod decode {
-        use quick_error::quick_error;
-
-        quick_error! {
-            #[derive(Debug)]
-            pub enum Error {
-                Corrupt(message: &'static str) {
-                    display("{}", message)
-                }
-                BitmapDecode{err: git_bitmap::ewah::decode::Error, kind: &'static str} {
-                    display("{} bitmap corrupt", kind)
-                    source(err)
-                }
-            }
-        }
-    }
-
-    pub fn decode(data: &[u8], object_hash: git_hash::Kind) -> Result<Link, decode::Error> {
-        let (id, data) = split_at_pos(data, object_hash.len_in_bytes())
-            .ok_or(decode::Error::Corrupt(
-                "link extension too short to read share index checksum",
-            ))
-            .map(|(id, d)| (git_hash::ObjectId::from(id), d))?;
-
-        if data.is_empty() {
-            return Ok(Link {
-                shared_index_checksum: id,
-                bitmaps: None,
-            });
-        }
-
-        let (delete, data) =
-            git_bitmap::ewah::decode(data).map_err(|err| decode::Error::BitmapDecode { kind: "delete", err })?;
-        let (replace, data) =
-            git_bitmap::ewah::decode(data).map_err(|err| decode::Error::BitmapDecode { kind: "replace", err })?;
-
-        if !data.is_empty() {
-            return Err(decode::Error::Corrupt("garbage trailing link extension"));
-        }
-
-        Ok(Link {
-            shared_index_checksum: id,
-            bitmaps: Some(Bitmaps { delete, replace }),
-        })
-    }
 }
